Right now it runs without global variables.
But this code is still far from good.
It needs refinement of variables. I've had to define two make two changes which at this time I feel were mistakes (instantiating Dec_Var with a fixed size column vector of size 54x1 in distributed_PF_V2.mlx, I think it should be only defined once the data is read, and changing Dec_Var to Dec_Var_1 in an if condition for 'ensuring the dimensions will match to store the voltages'.