Important change in branch KVL equation. Now results match with OpenDSS (are correct). Yay!
I was using branch indices for the voltages (the variables) when I actually intended to use the bus names as indices directly. Basically, for the KVL equation for the branch (i, j), I was essentially using voltages different from v_i and v_j, which has been rectified now. Still, there's a small difference in the PLoss values, and maybe that's an issue with the objective function or the NonLinear equations perhaps.