A patch to the `collectCentralizedInfo` function to get `sysInfo` to store the desired variables
Realized that some variables of `areaInfo` such as `PbattMax`, `Sder`, `Pmpp`, etc. do not correspond to what they should be corresponding to based on my latest definitions of `var_Full`, `var_0` and`var`. Tried rectifying it in `extractAreaInfo` but realized that then I'd have to change the optimizer variable names too. So for the sake of the meeting, have postponed this change in favour of a quick patch to `cCI.m`